+++ yt_dlp/__init__.py
@@ -20,7 +20,6 @@ from .cookies import SUPPORTED_BROWSERS, SUPPORTED_KEYRINGS
 from .downloader.external import get_external_downloader
 from .extractor import list_extractor_classes
 from .extractor.adobepass import MSO_INFO
-from .networking.utils import std_headers
 from .options import parseOpts
 from .postprocessor import (
     FFmpegExtractAudioPP,
@@ -62,6 +61,7 @@ from .utils import (
     variadic,
     write_string,
 )
+from .utils.networking import std_headers
 from .YoutubeDL import YoutubeDL
 
 _IN_CLI = False
