+++ yt_dlp/compat/__init__.py
@@ -75,7 +75,7 @@ else:
 def urllib_req_to_req(urllib_request):
     """Convert urllib Request to a networking Request"""
     from ..networking import Request
-    from ..networking.utils import HTTPHeaderDict
+    from ..utils.networking import HTTPHeaderDict
     return Request(
         urllib_request.get_full_url(), data=urllib_request.data, method=urllib_request.get_method(),
         headers=HTTPHeaderDict(urllib_request.headers, urllib_request.unredirected_hdrs),
